// src/generated/models/Afb_questionnairesectionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_questionnairesectionsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_questionnairesectionsstatecode = keyof typeof Afb_questionnairesectionsstatecode;
export const Afb_questionnairesectionsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_questionnairesectionsstatuscode = keyof typeof Afb_questionnairesectionsstatuscode;

export interface Afb_questionnairesectionsBase {
  afb_descriptionintroductive?: string;
  afb_intituleenanglais: string;
  afb_intituleenfrancais: string;
  afb_numerodordre: number;
  "afb_questionnaireassocie@odata.bind": string;
  afb_questionnairesectionid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_questionnairesectionsstatecode;
  statuscode?: Afb_questionnairesectionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_questionnairesections extends Afb_questionnairesectionsBase {
  afb_questionnaireassociename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_questionnaireassocie?: object;
  _afb_questionnaireassocie_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}