// src/generated/models/Afb_questionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_questionsafb_obligatoire = {
  0: 'Oui',
  747010001: 'Non'
} as const;
export type Afb_questionsafb_obligatoire = keyof typeof Afb_questionsafb_obligatoire;
export const Afb_questionsafb_piecejointerequise = {
  0: 'Oui',
  1: 'Non'
} as const;
export type Afb_questionsafb_piecejointerequise = keyof typeof Afb_questionsafb_piecejointerequise;
export const Afb_questionsafb_typedequestion = {
  0: 'Textecourt',
  1: 'Oui_Non',
  2: 'Pi_cejointe',
  3: 'Tableaudynamique',
  4: 'Num_rique',
  747010001: 'Textelong',
  747010002: 'Choixunique',
  747010003: 'Choixmultiples',
  747010004: 'Date',
  747010005: '_StatutC_PC_NC_NA'
} as const;
export type Afb_questionsafb_typedequestion = keyof typeof Afb_questionsafb_typedequestion;
export const Afb_questionsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_questionsstatecode = keyof typeof Afb_questionsstatecode;
export const Afb_questionsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_questionsstatuscode = keyof typeof Afb_questionsstatuscode;

export interface Afb_questionsBase {
  afb_aidealasaisie?: string;
  afb_codedelaquestion: string;
  afb_libelleenanglais: string;
  afb_libelleenfrancais: string;
  afb_numerodordre: number;
  afb_obligatoire: Afb_questionsafb_obligatoire;
  afb_piecejointerequise: Afb_questionsafb_piecejointerequise;
  afb_questionid: string;
  "afb_questionparente@odata.bind"?: string;
  afb_referencereglementaire?: string;
  "afb_section@odata.bind": string;
  afb_typedequestion: Afb_questionsafb_typedequestion;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_questionsstatecode;
  statuscode?: Afb_questionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_questions extends Afb_questionsBase {
  afb_obligatoirename?: string;
  afb_piecejointerequisename?: string;
  afb_questionparentename?: string;
  afb_sectionname?: string;
  afb_typedequestionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_questionparente?: object;
  _afb_questionparente_value?: string;
  afb_section?: object;
  _afb_section_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}